feat: add filter service for filtering locations
Adds a `FilterService` to filter locations by keyword. Updates mocks and tests to accommodate the new filtering logic.
The `fetchVPNLocations` in the mock now takes an `IPType` as parameter.
The `getRecentLocations` now no longer takes a keyword as a parameter.
The `getShownBanners` and `setShownBanners` method in the mock now return and accept list of `BannerType`.
The `isAvailable` property in the mock now returns `StoreState`.