feat: Add doOnListen stream extension and clear cached locations
Adds a `doOnListen` extension to the Stream class, allowing a callback function to be executed when the stream is listened to. This is used in the LocationsStore to trigger a refresh of locations when the stream is first listened to.

Also adds a "Clear cached locations" option to the QA Toolbox, which clears stored VPN location data from the local database.