Set manual code signing for iOS RunnerUITests
This commit changes the code signing style from `Automatic` to `Manual` for the `RunnerUITests` target in the iOS project. It also explicitly sets the `PROVISIONING_PROFILE_SPECIFIER` to "prod".

This change allows for more precise control over the code signing process, particularly for different build configurations and SDKs. The `DEVELOPMENT_TEAM` is now specified per SDK (iphoneos*), and the `CODE_SIGN_IDENTITY` is also set specifically for the `iphoneos*` SDK.